// src/lib/generated/billing/models/productCategory.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * kaizen-billing-and-subscription
 * Kaizen Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */

/**
 * Model ProductCategory
 */
export interface ProductCategory {
  /** @nullable */
  modifiedAt: string | null;
  createdAt: string;
  code: string;
  name: string;
  id: number;
}