// src/lib/generated/billing/models/updateProductCategoryDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * kaizen-billing-and-subscription
 * Kaizen Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateProductCategoryDto {
  name?: string;
  code?: string;
  linkedTopicIds?: number[];
  unlinkedTopicIds?: number[];
}